assetmanagementbackend/requirements.txt
asgiref==3.8.1
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.0
PyJWT==2.10.1
sqlparse==0.5.3
